refactor: rename app-sync to appsync
