+++ packages/amplify-graphql-directives/src/directives/appsync/aws-oidc.ts
