+++ packages/amplify-graphql-directives/src/directives/index.ts
@@ -1,5 +1,5 @@
 export type { Directive, DirectiveDefaults } from './directive';
-export * from './app-sync';
+export * from './appsync';
 export * from './v1';
 export { AuthDirective } from './auth';
 export { BelongsToDirective } from './belongs-to';
